+++ src/main/java/InterCoach/dto/RegisterRequest.java
@@ -1,11 +1,10 @@
 package InterCoach.dto;
 
-import com.fasterxml.jackson.annotation.JsonProperty;
 import jakarta.validation.constraints.Email;
 import jakarta.validation.constraints.NotBlank;
 import jakarta.validation.constraints.Size;
 
-public class UserRequest {
+public class RegisterRequest {
 
     @NotBlank
     @Size(min = 3, max = 50)
@@ -17,7 +16,6 @@ public class UserRequest {
 
     @NotBlank
     @Size(min = 8, max = 128)
-    @JsonProperty(access = JsonProperty.Access.WRITE_ONLY)
     private String password;
 
     public String getUsername() {